Refactor models in client module
